+++ examples/xmas/scripts/textGet.lua
@@ -1,0 +1,1 @@
+drawText(getLuaStore())